feat: address review feedback
