+++ js/src/domains/planning/index.tsx
@@ -7,7 +7,7 @@ import React, {
     useRef,
     useState,
 } from 'react';
-import { Box, Card, CardHeader } from '@mui/material';
+import { Card } from '@mui/material';
 import {
     LoadingSpinner,
     useRedirectToReplace,
@@ -34,14 +34,11 @@ import { MESSAGES } from '../messages';
 import { useDeleteScenario } from '../scenarios/hooks/useDeleteScenario';
 import { useGetScenario } from '../scenarios/hooks/useGetScenarios';
 import { useUpdateScenario } from '../scenarios/hooks/useUpdateScenario';
-import { BudgetSummary } from './components/budgeting/BudgetSummary';
 import { BudgetTable } from './components/budgeting/BudgetTable';
-import { CostPerDistrictSummary } from './components/budgeting/CostPerDistrictSummary';
-import { CostPerInterventionSummary } from './components/budgeting/CostPerInterventionSummary';
-import { PrevalenceSummary } from './components/budgeting/PrevalenceSummary';
 import { InterventionPlanHeader } from './components/interventionPlan/InterventionPlanHeader';
 import { InterventionPlanMap } from './components/interventionPlanMap/InterventionPlanMap';
 import { ScenarioRulesPanel } from './components/scenarioRule/ScenarioRulesPanel';
+import { ScenarioSummaryTab } from './components/ScenarioSummaryTab';
 import {
     PlanningProvider,
     usePlanningContext,
@@ -60,65 +57,6 @@ const styles = {
         height: '100%',
         display: 'flex',
         flexDirection: 'column',
-    },
-    summaryColumn: {
-        flex: 1,
-        width: '100%',
-        // Override PaperFullHeight's viewport height; flex fills MainColumn.
-        height: 0,
-        minHeight: 0,
-        maxHeight: '100%',
-        overflow: 'hidden',
-        display: 'flex',
-        flexDirection: 'column',
-    },
-    summaryHeaderCard: {
-        mb: 1,
-        flexShrink: 0,
-    },
-    // Equal py keeps the controls symmetric; minHeight 81px matches the other
-    // tabs' content box so they don't jump when switching.
-    summaryHeader: {
-        py: 2,
-        minHeight: '81px',
-    },
-    summaryGrid: {
-        flex: 1,
-        width: '100%',
-        minHeight: 0,
-        display: 'flex',
-        gap: 1,
-    },
-    summaryLeftColumn: {
-        flex: 4,
-        minWidth: 0,
-        minHeight: 0,
-        display: 'flex',
-        flexDirection: 'column',
-        gap: 1,
-    },
-    summaryRightColumn: {
-        flex: 8,
-        minWidth: 0,
-        minHeight: 0,
-        display: 'flex',
-        flexDirection: 'column',
-        gap: 1,
-    },
-    summaryWidget: {
-        flex: 1,
-        minHeight: 0,
-        overflow: 'hidden',
-    },
-    summaryWidgetThird: {
-        flex: 1,
-        minHeight: 0,
-        overflow: 'hidden',
-    },
-    summaryWidgetTwoThirds: {
-        flex: 2,
-        minHeight: 0,
-        overflow: 'hidden',
     },
 } satisfies SxStyles;
 
@@ -151,7 +89,9 @@ export const Planning: FC = () => {
     const { formatMessage } = useSafeIntl();
     const [activeTab, setActiveTab] = useState('map');
 
-    // Locked scenarios default to the summary tab, once, on first load.
+    // Locked scenarios default to the summary tab, but only on first load:
+    // the ref guard stops later scenario updates (e.g. toggling the lock, or a
+    // refetch) from yanking the user back to summary after they changed tabs.
     const didInitActiveTab = useRef(false);
     useEffect(() => {
         if (scenario && !didInitActiveTab.current) {
@@ -324,32 +264,7 @@ export const Planning: FC = () => {
                     />
                     <MainColumn>
                         {activeTab === 'summary' ? (
-                            <PaperFullHeight sx={styles.summaryColumn}>
-                                <Card sx={styles.summaryHeaderCard}>
-                                    <CardHeader
-                                        sx={styles.summaryHeader}
-                                        title={planHeader}
-                                    />
-                                </Card>
-                                <Box sx={styles.summaryGrid}>
-                                    <Box sx={styles.summaryLeftColumn}>
-                                        <Box sx={styles.summaryWidget}>
-                                            <BudgetSummary />
-                                        </Box>
-                                        <Box sx={styles.summaryWidget}>
-                                            <CostPerInterventionSummary />
-                                        </Box>
-                                    </Box>
-                                    <Box sx={styles.summaryRightColumn}>
-                                        <Box sx={styles.summaryWidgetThird}>
-                                            <PrevalenceSummary />
-                                        </Box>
-                                        <Box sx={styles.summaryWidgetTwoThirds}>
-                                            <CostPerDistrictSummary />
-                                        </Box>
-                                    </Box>
-                                </Box>
-                            </PaperFullHeight>
+                            <ScenarioSummaryTab header={planHeader} />
                         ) : (
                             <PaperFullHeight>
                                 <Card sx={styles.card}>
